More overloads and ddb context split
