+++ src/EfficientDynamoDb/Context/Operations/PutItem/IPutItemRequestBuilder.cs
@@ -1,6 +1,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 using EfficientDynamoDb.Context.FluentCondition.Core;
+using EfficientDynamoDb.DocumentModel;
 using EfficientDynamoDb.DocumentModel.ReturnDataFlags;
 
 namespace EfficientDynamoDb.Context.Operations.PutItem
@@ -20,8 +21,6 @@ namespace EfficientDynamoDb.Context.Operations.PutItem
     
     public interface IPutItemRequestBuilder<TEntity> where TEntity: class
     {
-        Task<PutItemEntityResponse<TEntity>> ExecuteAsync(CancellationToken cancellationToken = default);
-        
         IPutItemRequestBuilder<TEntity> WithReturnValues(ReturnValues returnValues);
         
         IPutItemRequestBuilder<TEntity> WithReturnConsumedCapacity(ReturnConsumedCapacity returnConsumedCapacity);
@@ -29,5 +28,15 @@ namespace EfficientDynamoDb.Context.Operations.PutItem
         IPutItemRequestBuilder<TEntity> WithReturnCollectionMetrics(ReturnItemCollectionMetrics returnItemCollectionMetrics);
         
         IPutItemRequestBuilder<TEntity> WithUpdateCondition(FilterBase condition);
+        
+        Task ExecuteAsync(CancellationToken cancellationToken = default);
+        
+        Task<TEntity?> ToEntityAsync(CancellationToken cancellationToken = default);
+        
+        Task<Document?> ToDocumentAsync(CancellationToken cancellationToken = default);
+        
+        Task<PutItemEntityResponse<TEntity>> ToEntityResponseAsync(CancellationToken cancellationToken = default);
+        
+        Task<PutItemEntityResponse<Document>> ToDocumentResponseAsync(CancellationToken cancellationToken = default);
     }
 }
