+++ src/EfficientDynamoDb/Context/Operations/UpdateItem/UpdateRequestBuilder.cs
@@ -7,6 +7,7 @@ using EfficientDynamoDb.Context.FluentCondition;
 using EfficientDynamoDb.Context.FluentCondition.Core;
 using EfficientDynamoDb.Context.FluentCondition.Factories;
 using EfficientDynamoDb.Context.Operations.Query;
+using EfficientDynamoDb.DocumentModel;
 using EfficientDynamoDb.DocumentModel.ReturnDataFlags;
 
 namespace EfficientDynamoDb.Context.Operations.UpdateItem
@@ -27,6 +28,32 @@ namespace EfficientDynamoDb.Context.Operations.UpdateItem
             _node = node;
         }
 
+        public Task ExecuteAsync(CancellationToken cancellationToken = default) => ToDocumentAsync(cancellationToken);
+
+        public async Task<TEntity?> ToEntityAsync(CancellationToken cancellationToken = default)
+        {
+            var classInfo = _context.Config.Metadata.GetOrAddClassInfo(typeof(TEntity));
+            return await _context.UpdateItemAsync<TEntity>(classInfo, _node, cancellationToken).ConfigureAwait(false);
+        }
+
+        public async Task<Document?> ToDocumentAsync(CancellationToken cancellationToken = default)
+        {
+            var classInfo = _context.Config.Metadata.GetOrAddClassInfo(typeof(TEntity));
+            return await _context.UpdateItemAsync<Document>(classInfo, _node, cancellationToken).ConfigureAwait(false);
+        }
+
+        public async Task<UpdateItemEntityResponse<TEntity>> ToEntityResponseAsync(CancellationToken cancellationToken = default)
+        {
+            var classInfo = _context.Config.Metadata.GetOrAddClassInfo(typeof(TEntity));
+            return await _context.UpdateItemResponseAsync<TEntity>(classInfo, _node, cancellationToken).ConfigureAwait(false);
+        }
+
+        public async Task<UpdateItemEntityResponse<Document>> ToDocumentResponseAsync(CancellationToken cancellationToken = default)
+        {
+            var classInfo = _context.Config.Metadata.GetOrAddClassInfo(typeof(TEntity));
+            return await _context.UpdateItemResponseAsync<Document>(classInfo, _node, cancellationToken).ConfigureAwait(false);
+        }
+
         public IAttributeUpdate<TEntity, TProperty> On<TProperty>(Expression<Func<TEntity, TProperty>> expression) => new AttributeUpdate<TEntity, TProperty>(this, expression);
 
         public IUpdateRequestBuilder<TEntity> WithReturnValues(ReturnValues returnValues) =>
@@ -43,9 +70,6 @@ namespace EfficientDynamoDb.Context.Operations.UpdateItem
 
         public IUpdateRequestBuilder<TEntity> WithUpdateCondition(Func<EntityFilter<TEntity>, FilterBase> filterSetup) =>
             new UpdateRequestBuilder<TEntity>(_context, new ConditionNode(filterSetup(Filter.ForEntity<TEntity>()), _node));
-
-        public async Task<UpdateItemEntityResponse<TEntity>> ExecuteAsync(CancellationToken cancellationToken = default) =>
-            await _context.UpdateItemAsync<TEntity>(_node, cancellationToken).ConfigureAwait(false);
 
         public IUpdateRequestBuilder<TEntity> WithPrimaryKey<TPk, TSk>(TPk pk, TSk sk) =>
             new UpdateRequestBuilder<TEntity>(_context, new PartitionAndSortKeyNode<TPk, TSk>(pk, sk, _node));
